+++ Assets/Scripts/Voxels/Generation/WorldGeneration.cs
@@ -11,9 +11,9 @@ namespace Voxels.Generation
     {
         public static ChunkData GenerateVoxels(ChunkData data, float noiseScale, int waterThreshold)
         {
-            int dirt = VoxelRegistry.GetId("std:Dirt");
-            int grass = VoxelRegistry.GetId("std:Grass");
-            int water = VoxelRegistry.GetId("std:Water");
+            ushort dirt = VoxelRegistry.GetId("std:Dirt");
+            ushort grass = VoxelRegistry.GetId("std:Grass");
+            ushort water = VoxelRegistry.GetId("std:Water");
             for (int x = 0; x < ChunkSize; x++)
             for (int z = 0; z < ChunkSize; z++)
             {
@@ -22,9 +22,9 @@ namespace Voxels.Generation
                 int groundPosition = Mathf.RoundToInt(noiseValue * ChunkHeight/2) + ChunkHeight/4;
                 for (int y = 0; y < ChunkHeight; y++)
                 {
-                    int voxelId = dirt;
+                    ushort voxelId = dirt;
                     if (y > groundPosition)
-                        voxelId = y < waterThreshold ? water : 0;
+                        voxelId = y < waterThreshold ? water : (ushort)0;
                     else if (y == groundPosition) voxelId = grass;
 
                     ChunkUtils.SetVoxel(data, new Vector3Int(x, y, z), voxelId);
